#37 Make AP GNN work for MAT (with low performance) (eng + fra)
